requirements.txt
boto3==1.34.0
mysql-connector-python==8.2.0
nltk==3.8.1
textblob==0.17.1
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
moto==4.2.14